v0.2 - Proper Season Setup!
Seasons are now properly setup before the start, giving the opportunity to choose which games to include and, in case of same racing series represented in different games, choosing which ones to include. This also means that finally seasons can be finished and new seasons started!
As a consequence of this, a lot of things had to be revised, leading in old savegames and old files not working anymore.
Additionally, a lot of other stuff has been added and changed and some vehicles and series were added.